+++ aa_forum/tests/test_templatetags.py
@@ -1091,7 +1091,7 @@ class TestForumDatetime(TestCase):
         expected_result = (
             f"{formatted_forum_date} "
             f'<sup>(<a href="{timezones_url}" target="_blank" rel="noopener noreferer" '
-            f'title="{link_title}" data-bs-tooltip="aa-forum-tooltip">'
+            f'title="{link_title}" data-bs-tooltip="aa-forum">'
             '<i class="fa-solid fa-circle-question"></i></a>)</sup>'
         )